Implement show range indicator.

+++ korge/src/commonMain/kotlin/com/xenotactic/korge/components/MutableShowRangeTimeComponent.kt
@@ -2,5 +2,5 @@ package com.xenotactic.korge.components
 
 data class MutableShowRangeTimeComponent(
     // Remaining time to show the range indicator in terms of milliseconds.
-    var showTimeRemainingMillis: Int
+    var showTimeRemainingMillis: Long
 )
